+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix: Improved support for Azure SQL Server Databases.
 - Update: SQL now uses native SQL script AI search instead of C# Assembly.
 - Update: UI, Templates and AI Avatar updated.
+- New: Create UI presets, like "Simple" or "Advanced".
 
 2024-06-16 v1.11.25